+++ classes/condition/passgrade/passgrade_condition.php
@@ -105,6 +105,7 @@ class passgrade_condition extends condition {
     /**
      * Saves the condition after it has been edited (or created)
      * @param object $formdata
+     * @return int The id of the saved condition record.
      */
     public function save_condition($formdata) {
         $cmid = clean_param($formdata->coursemodule ?? 0, PARAM_INT);
@@ -115,7 +116,7 @@ class passgrade_condition extends condition {
             'cmid' => $cmid,
         ];
 
-        $this->upsert($params, $formdata);
+        return $this->upsert($params, $formdata);
     }
 
     /**