+++ deletecondition.php
@@ -23,6 +23,7 @@
  */
 
 use local_coursedynamicrules\core\rule;
+use local_coursedynamicrules\helper\component_renderer;
 use local_coursedynamicrules\helper\rule_component_loader;
 
 require('../../config.php');
@@ -60,9 +61,9 @@ $condition = \local_coursedynamicrules\helper\ownership::get_condition($id, $cou
 $config = get_config('local_coursedynamicrules');
 
 $conditioninstance = rule_component_loader::create_condition_instance($condition, $courseid);
-$description = $conditioninstance->get_description();
+$description = component_renderer::escaped_description($conditioninstance);
 
-if ($delete === md5($config->confirmdeletecondition)) {
+if ($delete === md5($config->confirmdeletecondition ?? '')) {
     require_sesskey();
     // Delete condition.
     $conditioninstance->delete();
@@ -99,7 +100,7 @@ $continuebutton = new single_button(
     $continueurl,
     get_string('delete'),
     'post',
-    false,
+    single_button::BUTTON_SECONDARY,
     ['data-action' => 'delete']
 );
 echo $OUTPUT->confirm($message, $continuebutton, $conditionsurl);