+++ deleterule.php
@@ -56,7 +56,7 @@ $rule = \local_coursedynamicrules\helper\ownership::get_rule($id, $courseid);
 
 $config = get_config('local_coursedynamicrules');
 
-if ($delete === md5($config->confirmdeleterule)) {
+if ($delete === md5($config->confirmdeleterule ?? '')) {
     require_sesskey();
 
     // Delete rule.
@@ -95,7 +95,7 @@ $continuebutton = new single_button(
     $continueurl,
     get_string('delete'),
     'post',
-    false,
+    single_button::BUTTON_SECONDARY,
     ['data-action' => 'delete']
 );
 echo $OUTPUT->confirm($message, $continuebutton, $rulesurl);